changed double to float

+++ src/bvh_builder.cpp
@@ -33,7 +33,7 @@ typedef struct SplitNode SplitNode;
 the bounding box of the node should be split and location split_loc where the cut should be made.
 The function also uses bounding boxes of all the triangles (triangle_bounds), 
 positions of triangle centers (triangle_centers) and indices of triangles array (triangle_idxs).*/
-SplitNode get_split_node(const StackNode &node, double split_loc, int axis,
+SplitNode get_split_node(const StackNode &node, float split_loc, int axis,
                          const AABB *triangle_bounds, int *triangle_idxs, const Vector3 *triangle_centers) {
     // partition triangles in triangle_idxs[node.i0, node.i1] into two groups based on the position of their centers relative to split_loc
     auto it = std::partition(triangle_idxs + node.i0, triangle_idxs + node.i1,
@@ -69,13 +69,13 @@ SplitNode get_split_node(const StackNode &node, double split_loc, int axis,
 /*This function computes the cost of splitting the node along the given axis. 
 The node will be split in n_bins places along axis axis. For each split, the cost (surface area heuristic)
 will be computed and will be stored in costs*/
-void get_costs(const StackNode &node, double *costs, const AABB *triangle_bounds,
+void get_costs(const StackNode &node, float *costs, const AABB *triangle_bounds,
                 int *triangle_idxs, int n_bins, int axis, const Vector3 *triangle_centers){
 
     Vector3 diag = node.aabb.pmax - node.aabb.pmin;
     for(int i = 0; i < n_bins; i++){
         // compute the location of the split
-        double split_loc = node.aabb.pmin[axis] + diag[axis] * (i + 1) / (n_bins + 1);
+        float split_loc = node.aabb.pmin[axis] + diag[axis] * (i + 1) / (n_bins + 1);
         // get the two children of the node if it is split at split_loc
         SplitNode split_node = get_split_node(node, split_loc, axis, triangle_bounds, triangle_idxs, triangle_centers);
         // if the split is degenerate, set the cost to infinity
@@ -98,11 +98,11 @@ SplitNode split(const StackNode &node, const AABB *triangle_bounds,
     int axis = max_component(diag);
 
     // allocate memory for the costs and compute them
-    double *costs = new double[n_bins];
+    float *costs = new float[n_bins];
     get_costs(node, costs, triangle_bounds, triangle_idxs, n_bins, axis, triangle_centers);
 
     // choose the split with the smallest cost
-    double min_cost = infinity();
+    float min_cost = infinity();
     int split_bin = 0;
     for (int i = 0; i < n_bins; i++) {
         if (costs[i] < min_cost) {
@@ -113,7 +113,7 @@ SplitNode split(const StackNode &node, const AABB *triangle_bounds,
     delete[] costs;
 
     // compute the split
-    double split_loc = node.aabb.pmin[axis] + diag[axis] * (split_bin + 1) / (n_bins + 1);
+    float split_loc = node.aabb.pmin[axis] + diag[axis] * (split_bin + 1) / (n_bins + 1);
     return get_split_node(node, split_loc, axis, triangle_bounds, triangle_idxs, triangle_centers);
 }
